requirements.txt
# Core Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0

# Database and Caching
psycopg2-binary>=2.9.0  # PostgreSQL adapter
dj-database-url>=2.0.0  # Database URL parsing
django-redis>=5.2.0     # Redis caching
redis>=4.5.0            # Redis client

# Real-time Features (WebSocket)
channels>=4.0.0
channels-redis>=4.1.0

# File Storage and AWS
boto3>=1.26.0           # AWS SDK
django-storages>=1.13.0 # Cloud storage backends

# Background Tasks
celery>=5.2.0           # Task queue
django-celery-beat>=2.4.0  # Periodic tasks

# Authentication and Security
djangorestframework-authtoken>=1.0.0
django-oauth-toolkit>=1.7.0  # Optional: OAuth2 support
PyJWT>=2.6.0            # JWT tokens

# API Documentation (Optional)
drf-spectacular>=0.25.0  # OpenAPI 3 schema generation

# Filtering and Search
django-filter>=22.1

# Development and Testing
django-debug-toolbar>=4.0.0  # Development only
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.2.0      # Test data generation

# Utilities
python-decouple>=3.7    # Environment variable management
Pillow>=9.4.0          # Image processing
python-dateutil>=2.8.0 # Date utilities
requests>=2.28.0       # HTTP requests

# Production Server (Optional)
gunicorn>=20.1.0       # WSGI server
daphne>=4.0.0         # ASGI server for WebSocket support
whitenoise>=6.4.0     # Static file serving

# Monitoring and Logging (Optional)
sentry-sdk>=1.15.0    # Error tracking